Tiling avoids double draw, prefers overdraw
The image is divided into tiles. 
If there was a remainder that is less than the tile size, we would figure out the end coordinate of the image, and generate a "full" tile by insetting by the full tile size and masking to only the remainder. Now, we draw the tile as normal, i.e. the remainder's tile frame will go outside of the image, and Stable Diffusion will not draw in that portion.
This avoids cases where multile tiles and seams redrew areas, leading to distinct areas of the image being more detailed.